// Users Management/src/app/models/user.model.ts
import { Address } from "./address.model";
import { Company } from "./company.model";

export interface User {
    id?: number;
    name: string;
    username: string;
    email: string;
    address?: Address;
    phone: string;
    website: string;
    company?: Company;
    age?: number;
    dateOfBirth?: Date;
}